// mobile/convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  agent: typeof agent;
  ai: typeof ai;
  ai_config: typeof ai_config;
  auth: typeof auth;
  campaigns: typeof campaigns;
  chat: typeof chat;
  contacts: typeof contacts;
  files: typeof files;
  filesInternal: typeof filesInternal;
  http: typeof http;
  index: typeof index;
  integrations: typeof integrations;
  messages: typeof messages;
  notifications: typeof notifications;
  orders: typeof orders;
  products: typeof products;
  salla: typeof salla;
  seed: typeof seed;
  stats: typeof stats;
  templates: typeof templates;
  users: typeof users;
  webhookEvents: typeof webhookEvents;
  whatsapp: typeof whatsapp;
  workflows: typeof workflows;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;

/**
 * A utility for referencing Convex functions in your app's internal API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = internal.myModule.myFunction;
 * ```
 */
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;